tests/wpt/tests/html/semantics/forms/the-textarea-element/textarea-setcustomvalidity.html

<!DOCTYPE HTML>
<title>textarea setCustomValidity</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<textarea id='textarea_test'></textarea>

<script>

test(() => {
  let elem = document.getElementById("textarea_test");
  assert_false(elem.validity.customError);
  elem.setCustomValidity("custom error");
  assert_true(elem.validity.customError);
}, "textarea setCustomValidity is correct")

</script>
